+++ apps/perf-test/src/scenarios/ThemeProvider.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import { ThemeProvider } from '@fluentui/react-next';
+
+const scenario = <ThemeProvider />;
+
+export default scenario;
