+++ packages/react-next/.storybook/decorators/ThemeProvider.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { ThemeProvider } from '@fluentui/react-next';
+import { useCustomizationOptions } from '../knobs/theme';
+
+export const ThemeProviderDecorator = storyFn => {
+  const customizationOptions = useCustomizationOptions();
+  const { customizations, isDark } = customizationOptions;
+  const themeProviderProps = {
+    theme: customizations ? { tokens: customizations.settings.theme } : undefined,
+    style: {
+      background: isDark ? 'black' : undefined,
+    },
+  };
+
+  return <ThemeProvider {...themeProviderProps}>{storyFn()}</ThemeProvider>;
+};
