+++ packages/react-next/.storybook/preview.js
@@ -7,10 +7,12 @@ import { withPerformance } from 'storybook-addon-performance';
 import { withKnobs } from '@storybook/addon-knobs';
 import { useCustomizationOptions } from './knobs/theme';
 import { ThemeProvider } from '@fluentui/react-next';
+import { ThemeProviderDecorator } from './decorators/ThemeProvider';
 
 addDecorator(withA11y());
 addDecorator(withPerformance);
 addDecorator(withKnobs({ escapeHTML: false }));
+addDecorator(ThemeProviderDecorator);
 addParameters({
   a11y: {
     manual: true,
@@ -33,21 +35,7 @@ function loadStories() {
     Object.keys(story).forEach(exampleName => {
       const example = story[exampleName];
       if (typeof example === 'function') {
-        story[exampleName] = () => {
-          const customizationOptions = useCustomizationOptions();
-          const { customizations, isDark } = customizationOptions;
-          const themeProviderProps = {
-            theme: customizations ? { tokens: customizations.settings.theme } : undefined,
-            style: {
-              background: isDark ? 'black' : undefined,
-            },
-          };
-
-          return React.createElement(ThemeProvider, {
-            ...themeProviderProps,
-            children: example(),
-          });
-        };
+        story[exampleName] = () => example();
       }
     });
   }
